pyproject.toml
[tool.poetry]
name = "babies-data-processor"
version = "0.1.0"
description = ""
authors = ["A Z <madzul@madz-n-cya.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.3.3"
numpy = "^2.3.3"
mysql-connector-python = "^9.4.0"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
